refactor: clear form after submit or cancel in edit api key form
